Use HashMap to retrieve values using JNI
Instead of directly creating an AudioFile in the JNI from taglib's
PropertyMap, it is better to convert it into a java/kotlin HashMap.
The HashMap can be mapped in the Kotlin code to produce an AudioFile.
This way any new tag fields can be supported or removed by only changing
the Kotlin code. This is beneficial as the JNI code does not provide any
compile time safety. Errors is the JNI code will lead to runtime crashes
whereas the compiler will catch a large number of errors at the compile
time.